Rename copy() to __copy__() in Python version of xml.etree.ElementTree.
The C version defines __copy__() and __deepcopy() while the Python
version defines copy(). The documentation does not mention copy(), so
it is presumed to be in error. This unifies the definition of __copy__()
between the Python and C version, making the proper way to copy an
Element be copy.copy(ET.Element("foo")).